[API] Enable customer registration using new metadata
