+++ disabledTests/Api/Shop/CustomersTest.php
@@ -44,49 +44,6 @@ final class CustomersTest extends JsonApiTestCase
     }
 
     /** @test */
-    public function it_registers_customers(): void
-    {
-        $this->loadFixturesFromFiles(['channel.yaml']);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/shop/customers',
-            server: self::CONTENT_TYPE_HEADER,
-            content: json_encode([
-                'firstName' => 'John',
-                'lastName' => 'Doe',
-                'email' => 'shop@example.com',
-                'password' => 'sylius',
-                'subscribedToNewsletter' => true,
-            ], JSON_THROW_ON_ERROR),
-        );
-
-        $response = $this->client->getResponse();
-
-        $this->assertResponseCode($response, Response::HTTP_NO_CONTENT);
-    }
-
-    /** @test */
-    public function it_allows_customer_to_log_in(): void
-    {
-        $this->loadFixturesFromFiles(['authentication/customer.yaml']);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/shop/authentication-token',
-            server: self::CONTENT_TYPE_HEADER,
-            content: json_encode([
-                'email' => 'oliver@doe.com',
-                'password' => 'sylius'
-            ], JSON_THROW_ON_ERROR),
-        );
-
-        $response = $this->client->getResponse();
-
-        $this->assertResponse($response, 'shop/log_in_customer_response', Response::HTTP_OK);
-    }
-
-    /** @test */
     public function it_allows_customer_to_update_its_data(): void
     {
         $loadedData = $this->loadFixturesFromFiles(['authentication/customer.yaml']);
