+++ features/account/verification/verifying_email_address.feature
@@ -45,14 +45,14 @@ Feature: Verifying account's email address
         And I have already verified my account
         Then I should not be able to resend the verification email
 
-    @ui @email #api / Customer resource needed
+    @ui @email @api
     Scenario: Receiving account verification email after registration
         When I register with email "ghastly@bespoke.com" and password "suitsarelife"
         Then I should be notified that my account has been created and the verification email has been sent
         And 1 email should be sent to "ghastly@bespoke.com"
         But I should not be able to log in as "ghastly@bespoke.com" with "suitsarelife" password
 
-    @ui @email #api / Customer resource needed
+    @ui @email @api
     Scenario: Do not send verification email when account verification on the channel is not required
         Given "United States" channel has account verification disabled
         When I register with email "ghastly@bespoke.com" and password "suitsarelife"
