+++ src/Sylius/Behat/Resources/config/suites.yml
@@ -4,8 +4,7 @@
 imports:
 #    - suites/api/account/address_book.yml
 #    - suites/api/account/customer.yml
-#    - suites/api/account/customer_registration.yml
-#    - suites/api/account/email_verification.yml
+    - suites/api/account/customer_registration.yml
     - suites/api/account/email_verification.yml
     - suites/api/account/login.yaml
     - suites/api/addressing/managing_countries.yaml
