+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/command_handlers.xml
@@ -18,16 +18,16 @@
     <services>
         <defaults public="true" />
 
-<!--        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Account\RegisterShopUserHandler">-->
-<!--            <argument type="service" id="sylius.factory.shop_user" />-->
-<!--            <argument type="service" id="sylius.manager.shop_user" />-->
-<!--            <argument type="service" id="Sylius\Bundle\CoreBundle\Resolver\CustomerResolverInterface" />-->
-<!--            <argument type="service" id="sylius.repository.channel" />-->
-<!--            <argument type="service" id="sylius.shop_user.token_generator.email_verification" />-->
-<!--            <argument type="service" id="sylius.command_bus" />-->
-<!--            <tag name="messenger.message_handler" bus="sylius.command_bus" />-->
-<!--            <tag name="messenger.message_handler" bus="sylius_default.bus" />-->
-<!--        </service>-->
+        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Account\RegisterShopUserHandler">
+            <argument type="service" id="sylius.factory.shop_user" />
+            <argument type="service" id="sylius.manager.shop_user" />
+            <argument type="service" id="Sylius\Bundle\CoreBundle\Resolver\CustomerResolverInterface" />
+            <argument type="service" id="sylius.repository.channel" />
+            <argument type="service" id="sylius.shop_user.token_generator.email_verification" />
+            <argument type="service" id="sylius.command_bus" />
+            <tag name="messenger.message_handler" bus="sylius.command_bus" />
+            <tag name="messenger.message_handler" bus="sylius_default.bus" />
+        </service>
 
 <!--        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Cart\PickupCartHandler">-->
 <!--            <argument type="service" id="sylius.factory.order" />-->
@@ -172,21 +172,21 @@
             <tag name="messenger.message_handler" bus="sylius_default.bus" />
         </service>
 
-<!--        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Account\SendAccountRegistrationEmailHandler">-->
-<!--            <argument type="service" id="sylius.repository.shop_user" />-->
-<!--            <argument type="service" id="sylius.repository.channel" />-->
-<!--            <argument type="service" id="sylius.email_sender" />-->
-<!--            <tag name="messenger.message_handler" bus="sylius.command_bus" />-->
-<!--            <tag name="messenger.message_handler" bus="sylius_default.bus" />-->
-<!--        </service>-->
+        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Account\SendAccountRegistrationEmailHandler">
+            <argument type="service" id="sylius.repository.shop_user" />
+            <argument type="service" id="sylius.repository.channel" />
+            <argument type="service" id="sylius.email_sender" />
+            <tag name="messenger.message_handler" bus="sylius.command_bus" />
+            <tag name="messenger.message_handler" bus="sylius_default.bus" />
+        </service>
 
-<!--        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Account\SendAccountVerificationEmailHandler">-->
-<!--            <argument type="service" id="sylius.repository.shop_user" />-->
-<!--            <argument type="service" id="sylius.repository.channel" />-->
-<!--            <argument type="service" id="sylius.email_sender" />-->
-<!--            <tag name="messenger.message_handler" bus="sylius.command_bus" />-->
-<!--            <tag name="messenger.message_handler" bus="sylius_default.bus" />-->
-<!--        </service>-->
+        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Account\SendAccountVerificationEmailHandler">
+            <argument type="service" id="sylius.repository.shop_user" />
+            <argument type="service" id="sylius.repository.channel" />
+            <argument type="service" id="sylius.email_sender" />
+            <tag name="messenger.message_handler" bus="sylius.command_bus" />
+            <tag name="messenger.message_handler" bus="sylius_default.bus" />
+        </service>
 
 <!--        <service id="Sylius\Bundle\ApiBundle\CommandHandler\Checkout\SendOrderConfirmationHandler">-->
 <!--            <argument type="service" id="sylius.email_sender" />-->
